+++ examples/node/helloworld/firedalerts_async.js
@@ -25,7 +25,7 @@ exports.main = function(opts, callback) {
     opts = opts || {};
     
     var username = opts.username    || "admin";
-    var password = opts.password    || "changeme";
+    var password = opts.password    || "1";
     var scheme   = opts.scheme      || "https";
     var host     = opts.host        || "localhost";
     var port     = opts.port        || "8089";
@@ -56,21 +56,23 @@ exports.main = function(opts, callback) {
             },
             // Print them out
             function(firedAlertGroups, done) {
-                // Get the list of all alert, including the all group (represented by "-")
-                var alertGroups = firedAlertGroups.list();
+                // Get the list of all fired alert groups, including the all group (represented by "-")
+                var firedAlertGroups = firedAlertGroups.list();
 
                 console.log("Fired alert groups:");
                 Async.seriesEach(
-                    alertGroups,
-                    function(group, index, seriescallback) {
-                        group.list(function(err, jobs, alertGroup){
-                            // How many search jobs fired this alert?
-                            console.log(group.name, "(Count:", group.count(), ")");
-                            // Print the properties for each job that fired this alert (default of 30 per alert)
-                            for(var i = 0; i < jobs.length; i++) {
-                                var job = jobs[i];
-                                for (var key in job.properties()) {
-                                    console.log(key + ":", job.properties()[key]);
+                    firedAlertGroups,
+                    function(firedAlertGroup, index, seriescallback) {
+                        firedAlertGroup.list(function(err, firedalerts, alertGroup){
+                            // How many times was this alert fired?
+                            console.log(firedAlertGroup.name, "(Count:", firedAlertGroup.count(), ")");
+                            // Print the properties for each fired alert (default of 30 per alert group)
+                            for(var i = 0; i < firedalerts.length; i++) {
+                                var firedAlert = firedalerts[i];
+                                for (var key in firedAlert.properties()) {
+                                    if (firedAlert.properties().hasOwnProperty(key)) {
+                                        console.log("\t", key, ":", firedAlert.properties()[key]);
+                                    }
                                 }
                                 console.log();
                             }